requirements.txt
discord.py==1.7.3
python-dotenv==0.19.0
discord-py-slash-command==3.0.1
discord-together==1.1.1